==> scripts/2 prepare data based on family level.R <==
loadObj(fsdata, "data/stroke_dat_1201.rds")
## stroke patients and controls
# stroke_data <- fsdata %>% dplyr::filter(!is.na(family_no_isstroke))
# write_excel_csv(stroke_data,"data\\stk_dat_1112_utf8.csv")
## get all SNPs
SNP_names <- getSNPName(fsdata)

## get Summary of the SNPs
# allele frequency
Al_Freq_table <- map_dfr(SNP_names, 
                         getAlleleFrequency, 
                         data = fsdata, 
                         drop.freq = FALSE)
Al_Freq_table <- Al_Freq_table %>% dplyr::select(rsid, A,T,G,C,A_N,T_N,G_N,C_N)
write_excel_csv(Al_Freq_table, "results\\Allele_Freq.csv")
# genotype frquency
geno_freq_table <- map_dfr(SNP_names, genoFreq, data = fsdata)
write_excel_csv(geno_freq_table, "results\\genotype_Freq.csv")
# chunk SNP with calling rate > 0.5
SNP_keep <- getVariable(geno_freq_table, rsid, miss_freq < 0.5) %>% unique()
SNP_remove <- setdiff(SNP_names, SNP_keep)

## remove imcomplete samples
fsdata <- fsdata %>% 
  dplyr::select(-one_of(SNP_remove)) %>%
  mutate(call_num = rowSums(is.na(.[SNP_keep])))
sample.delete <- fsdata %>% 
  filter(call_num > 100 |is.null(isfamily_no)|is.na(stroke_proven)|!stroke_proven %in% c(0,1))

## delete incompete families
stk_dat_comp <- fsdata %>% 
  dplyr::filter(! pid %in% sample.delete$pid) %>%
  # fix erronous coding
  filter(!family_no %in% c("1404113","1218") & !pid %in% c("3703", "142110704","142110705","170911702","170911703")) %>%
  mutate(memid = ifelse(pid == "141411002",4,memid),
         memid = ifelse(pid == "141411003",5,memid),
         memid = ifelse(pid == "170910602",4,memid)) %>%
  # calculate probands
  mutate(id_aff = ifelse(stroke_proven == 1 & !memid %in% c(2,3), memid, 99),
         proband_err = ifelse(memid == 1 & stroke_proven != 1, 1, 0),
         pid = as.character(pid)) %>%
  group_by(family_no) %>%
  mutate(proband = min(id_aff, na.rm = TRUE),
         memid = ifelse(proband_err == 1, max(memid)+1, memid),
         memid = ifelse(!is.na(proband) & memid !=1 & memid == proband, 1, memid),
         n_fam = n()) %>%
  filter(n_fam >= 2 & proband != 99) %>%
    # delete unrelated variables
  dplyr::select(-id_aff,-proband, -proband_err, -n_fam) %>%
  ungroup()

# genotype freq after remove
geno_freq_table <- map_dfr(SNP_keep, genoFreq, data = stk_dat_comp) %>% arrange(miss_freq)

## relativeship
M <- map_dfc(stk_dat_comp[SNP_keep], geno_mat) %>% as.matrix()

## add parentid to data
stk_dat_comp <- addPid(stk_dat_comp, pid = pid, famid = family_no, memid = memid)
## diagnose hypertension and diabetes
loadObj(stk_dat_comp, "data\\stk_dat.rds")

## output to data
write_rds(stk_dat_comp,"data/stk_dat.rds")
