==> scripts/5 analysis 1.R <==
loadObj(stk_dat, "data/stk_dat.rds")
loadObj(fsdata,"data/stroke_dat_1201.rds")
library(data.table)
## discriptive family information
fsdata_summary <- setDT(fsdata)[,.(sib_num = sum(memid %in% c(1,4,5,6,7,8,9)), 
                            fa_num = sum(memid == 2), 
                            mo_num = sum(memid == 3)), family_no][,.(sib_num,fa_num,mo_num,pa_num = fa_num+mo_num)]

table1 <- fsdata_summary[,.(sum_pop = .N, sum_fa = sum(fa_num), sum_mo = sum(mo_num), sum_pa = sum(pa_num ==2)), sib_num] %>%
  arrange(sib_num)
## association test

# conditional logistic regression
snps <- colnames(stk_dat) %>% str_subset("^rs")
stk_add <- map(stk_dat[snp], )
cor_mat <- cor(stk_dat[snps])

library(survival)
clogit(is_stroke ~ rs2043211 + strata(fid), data = stk_dat,method = "exact") %>% summary()
